requirements.txt
tk
Pillow
hijri-converter
requests
pytz
